requirements.txt
streamlit==1.30.0
pillow==10.1.0
numpy==1.26.1
opencv-python-headless==4.8.1.78
imageio==2.31.1